+++ client/src/io/worxbend/gitea4s/http/ContentsParams.scala
@@ -1,0 +1,8 @@
+package io.worxbend.gitea4s.http
+
+final case class ContentsParams(
+    ref: Option[String] = None
+)
+
+object ContentsParams:
+  val default: ContentsParams = ContentsParams()